+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Support for references and acknowledgements on flaw creation (`OSIDB-2319`)
 * Sort Advanced Search Options alphabetically (`OSIDB-2805`)
 * Add tracker links on affects and flaw form (`OSIDB-2630`)
+* Add button to expand all affects (`OSIDB-2817`)
 * Sort affects by product family, alphabetically (`OSIDB-2533`)
 * Sort impact options logically (`OSIDB-2791`)
 * Advanced search on emptiness for CVE ID (`OSIDB-2806`)
@@ -33,6 +34,7 @@
 ### Removed
 * Removed is_major_incident usage (`OSIDB-2778`)
 * Removed comment.type usage (`OSIDB-2781`)
+* Removed Flaw.meta usage (`OSIDB-2801`)
 
 ## [Unreleased] (TODO update for 2024.1.1 and 2024.2.0 releases)